+++ packages/api-headless-cms/src/content/plugins/schema/createReadSDL.ts
@@ -26,7 +26,7 @@ export const createReadSDL: CreateManageSDL = ({ model, fieldTypePlugins }): str
     const fieldsRender = renderFields({ model, type: "read", fieldTypePlugins });
 
     return `
-        """${model.description}"""
+        """${model.description || ""}"""
         type ${rTypeName} {
             id: ID
             createdOn: DateTime