+++ packages/app-page-builder/src/editor/recoil/actions/saveRevision/types.ts
@@ -1,4 +1,5 @@
 export type SaveRevisionActionArgsType = {
+    debounce?: boolean;
     onFinish?: () => void;
 };